+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
 
-## Beta 1.6 (tba)
+## Beta 1.6 (13.01.2020)
 
 ### Added
 
 - Added "Temporarily Disable Physics Toggling" button in the settings. If your vehicle gets stuck in the air, you can temporarily disable physics toggling. Physics toggling will be enabled back next time you restart the game
 - Added the date of report generation to mod report
+- Strawberry field now is disabled when player is far away from it
+- Piers and railroad tunnels are also disabled
 
 ### Changes
 
 - Changed the initialization delay to 2 seconds
+- Code optimization and enchancements
 
 ### Bug Fixes